# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: StockBTP - API de gestion de stock pour entreprise de construction de conteneurs
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stock-movements:
    get:
      operationId: listStockMovements
      tags: [stock-movements]
      summary: Historique des mouvements
      parameters:
        - name: product_id
          in: query
          schema: { type: integer }
        - name: project_id
          in: query
          schema: { type: integer }
        - name: type
          in: query
          schema: { type: string, enum: [IN, OUT] }
        - name: from_date
          in: query
          schema: { type: string }
        - name: to_date
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: Liste des mouvements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockMovement"
    post:
      operationId: createStockMovement
      tags: [stock-movements]
      summary: Enregistrer un mouvement de stock
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockMovementInput"
      responses:
        "201":
          description: Mouvement créé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockMovement"
        "400":
          description: Stock insuffisant ou données invalides
  /stock-movements/{id}:
    get:
      operationId: getStockMovement
      tags: [stock-movements]
      summary: Détails d'un mouvement
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Mouvement trouvé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockMovement"

  # ─── PROJECTS ────────────────────────────────────────────────────────────────
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: Liste des utilisateurs
      responses:
        "200":
          description: Liste des utilisateurs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Créer un utilisateur
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Utilisateur créé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Détails d'un utilisateur
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Utilisateur trouvé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Modifier un utilisateur
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Utilisateur modifié
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ─── AUTH ────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    StockMovement:
      type: object
      required: [id, productId, type, quantity, reason, createdById, createdAt]
      properties:
        id: { type: integer }
        productId: { type: integer }
        productName: { type: ["string", "null"] }
        type: { type: string, enum: [IN, OUT] }
        quantity: { type: integer }
        reason: { type: string }
        projectId: { type: ["integer", "null"] }
        projectName: { type: ["string", "null"] }
        createdById: { type: integer }
        createdByName: { type: ["string", "null"] }
        createdAt: { type: string }
    StockMovementInput:
      type: object
      required: [productId, type, quantity, reason, createdById]
      properties:
        productId: { type: integer }
        type: { type: string, enum: [IN, OUT] }
        quantity: { type: integer, minimum: 1 }
        reason: { type: string, minLength: 1 }
        projectId: { type: ["integer", "null"] }
        createdById: { type: integer }

    # ─── PROJECT ─────────────────────────────────────────────────────────────
...
    User:
      type: object
      required: [id, fullName, email, role, createdAt]
      properties:
        id: { type: integer }
        fullName: { type: string }
        email: { type: string }
        role: { type: string, enum: [admin, manager, worker] }
        createdAt: { type: string }
    UserInput:
      type: object
      required: [fullName, email, role, password]
      properties:
        fullName: { type: string, minLength: 1 }
        email: { type: string }
        role: { type: string, enum: [admin, manager, worker] }
        password: { type: string, minLength: 6 }
    UserUpdate:
      type: object
      properties:
        fullName: { type: string }
        role: { type: string, enum: [admin, manager, worker] }
...