# lib/api-zod/src/generated/types/stockMovement.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * StockBTP - API de gestion de stock pour entreprise de construction de conteneurs
 * OpenAPI spec version: 0.1.0
 */
import type { StockMovementType } from './stockMovementType';

export interface StockMovement {
  id: number;
  productId: number;
  /** @nullable */
  productName?: string | null;
  type: StockMovementType;
  quantity: number;
  reason: string;
  /** @nullable */
  projectId?: number | null;
  /** @nullable */
  projectName?: string | null;
  createdById: number;
  /** @nullable */
  createdByName?: string | null;
  createdAt: string;
}